Per-channel convo history, support Gemma, Mistral and OpenRouter vision
Change paradigm of conversation history to not be per-user across all channels,
but just history of the channel where the bot was invoked, be it DM or normal
channel. Typing /reset command for conversation will reset the history on
that particular channel. Use this when the bot gets stuck in a hallucination loop.

Add support for Gemma on OpenRouter and FireLlava vision on OpenRouter.

Handle async properly which seems to fix stability.